# spring-server/src/main/java/io/swagger/api/CalculatorApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.4.5).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.swagger.api;

import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;

import io.swagger.annotations.Api;
import io.swagger.annotations.ApiOperation;
import io.swagger.annotations.ApiParam;
import io.swagger.annotations.ApiResponse;
import io.swagger.annotations.ApiResponses;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2019-06-25T10:41:00.449Z")

@Api(value = "calculator", description = "the calculator API")
public interface CalculatorApi {

    @ApiOperation(value = "Gaussian", nickname = "compute", notes = "calcola i parametri delle gaussiane che descrivono la mappa", response = String.class, tags={ "calculator", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Calcolo svolto correttamente!", response = String.class),
        @ApiResponse(code = 400, message = "Operazione non valida!") })
    @RequestMapping(value = "/calculator/{latitude}/{longitude}/{altitude}",
        produces = { "text/plain" }, 
        method = RequestMethod.GET)
    ResponseEntity<String> compute(@ApiParam(value = "",required=true) @RequestHeader("latitude") String latitude,
    								@ApiParam(value = "",required=true) @RequestHeader("longitude") String longitude,
    								@ApiParam(value = "",required=true) @RequestHeader("altitude") String altidude);

}